+++ src/main/java/com/com2here/com2hereback/controller/TokenController.java
@@ -4,7 +4,7 @@ import com.com2here.com2hereback.common.BaseException;
 import com.com2here.com2hereback.common.BaseResponseStatus;
 import com.com2here.com2hereback.common.CMResponse;
 import com.com2here.com2hereback.dto.UserTokenResponseDto;
-import com.com2here.com2hereback.service.UserService;
+import com.com2here.com2hereback.service.TokenService;
 import com.com2here.com2hereback.vo.UserTokenResponseVo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,12 +16,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/token")
 public class TokenController {
 
-    private final UserService userService;
+    private final TokenService tokenService;
 
     @GetMapping("/rotate")
     public CMResponse<UserTokenResponseVo> rotateToken() {
         try{
-            UserTokenResponseDto userTokenResponseDto = userService.rotateToken();
+            UserTokenResponseDto userTokenResponseDto = tokenService.rotateToken();
             UserTokenResponseVo userTokenResponseVo = UserTokenResponseVo.dtoToVo(userTokenResponseDto);
             return CMResponse.success(BaseResponseStatus.SUCCESS, userTokenResponseVo);
         } catch (BaseException e) {
