+++ src/main/java/com/com2here/com2hereback/service/TokenService.java
@@ -1,0 +1,7 @@
+package com.com2here.com2hereback.service;
+
+import com.com2here.com2hereback.dto.UserTokenResponseDto;
+
+public interface TokenService {
+    UserTokenResponseDto rotateToken();
+}
