+++ src/main/java/com/com2here/com2hereback/service/UserService.java
@@ -4,7 +4,6 @@ import com.com2here.com2hereback.dto.ChgPasswordRequestDto;
 import com.com2here.com2hereback.dto.ShowUserResponseDto;
 import com.com2here.com2hereback.dto.UserLoginResponseDto;
 import com.com2here.com2hereback.dto.UserRequestDto;
-import com.com2here.com2hereback.dto.UserTokenResponseDto;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface UserService {
@@ -15,6 +14,5 @@ public interface UserService {
     void updateUser(String nickname, String email, MultipartFile profileImage);
     void deleteUser(UserRequestDto userRequestDto);
     void chgPassword(ChgPasswordRequestDto chgPasswordRequestDto);
-    UserTokenResponseDto rotateToken();
     UserLoginResponseDto registerOrLoginSocialUser(String email, String nickname, String provider, String oauthId, String profileImageUrl);
 }
